refactor: Modules list refactor (#3760)
* Separating components into separate files

* types added

* Types have been corrected

* Reducing identical delete boxes

* PR corrections
