+++ src/resources/other/kymaModules.routes.js
@@ -1,17 +1,6 @@
-import {
-  Button,
-  CheckBox,
-  FlexibleColumnLayout,
-  List,
-  ListItemStandard,
-  MessageStrip,
-  Text,
-} from '@ui5/webcomponents-react';
-
-import React, { Suspense, useEffect, useMemo, useState } from 'react';
-import { Route, useNavigate, useParams } from 'react-router-dom';
-
-import pluralize from 'pluralize';
+import { Button, FlexibleColumnLayout } from '@ui5/webcomponents-react';
+import React, { Suspense, useEffect, useState } from 'react';
+import { Route, useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import { ErrorBoundary } from 'shared/components/ErrorBoundary/ErrorBoundary';
 import { ResourceCreate } from 'shared/components/ResourceCreate/ResourceCreate';
@@ -22,7 +11,6 @@ import ExtensibilityDetails from 'components/Extensibility/ExtensibilityDetails'
 import { t } from 'i18next';
 import { createPortal } from 'react-dom';
 import { useDeleteResource } from 'shared/hooks/useDeleteResource';
-import { useDelete } from 'shared/hooks/BackendAPI/useMutation';
 import { useNotification } from 'shared/contexts/NotificationContext';
 import { useCreateResource } from 'shared/ResourceForm/useCreateResource';
 import { cloneDeep } from 'lodash';
@@ -30,18 +18,7 @@ import {
   useKymaQuery,
   useModuleTemplatesQuery,
 } from 'components/KymaModules/kymaModulesQueries';
-import { useGetScope, useSingleGet } from 'shared/hooks/BackendAPI/useGet';
-
-import {
-  checkIfAssociatedResourceLeft,
-  deleteAssociatedResources,
-  deleteCrResources,
-  fetchResourceCounts,
-  generateAssociatedResourcesUrls,
-  getAssociatedResources,
-  getCRResource,
-  handleItemClick,
-} from 'components/KymaModules/deleteModulesHelpers';
+import { ModulesDeleteBox } from 'components/KymaModules/components/ModulesDeleteBox';
 
 const KymaModulesList = React.lazy(() =>
   import('../../components/KymaModules/KymaModulesList'),
@@ -137,73 +114,6 @@ const ColumnWraper = ({ defaultColumn = 'list', namespaced = false }) => {
     skip: !(layoutState?.midColumn?.resourceName || resourceName),
   });
 
-  const fetchFn = useSingleGet();
-  const getScope = useGetScope();
-  const navigate = useNavigate();
-  const deleteResourceMutation = useDelete();
-
-  const [resourceCounts, setResourceCounts] = useState({});
-  const [forceDeleteUrls, setForceDeleteUrls] = useState([]);
-  const [crUrls, setCrUrls] = useState([]);
-  const [allowForceDelete, setAllowForceDelete] = useState(false);
-  const [associatedResourceLeft, setAssociatedResourceLeft] = useState(false);
-
-  const associatedResources = useMemo(
-    () =>
-      getAssociatedResources(
-        openedModuleIndex,
-        activeKymaModules,
-        kymaResource,
-        moduleTemplates,
-      ),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [openedModuleIndex, moduleTemplates],
-  );
-
-  useEffect(() => {
-    const fetchCounts = async () => {
-      const counts = await fetchResourceCounts(associatedResources, fetchFn);
-      const urls = await generateAssociatedResourcesUrls(
-        associatedResources,
-        fetchFn,
-        clusterUrl,
-        getScope,
-        namespaceUrl,
-        navigate,
-      );
-      const crUResources = getCRResource(
-        openedModuleIndex,
-        activeKymaModules,
-        kymaResource,
-        moduleTemplates,
-      );
-      const crUrl = await generateAssociatedResourcesUrls(
-        crUResources,
-        fetchFn,
-        clusterUrl,
-        getScope,
-        namespaceUrl,
-        navigate,
-      );
-      setResourceCounts(counts);
-      setForceDeleteUrls(urls);
-      setCrUrls([crUrl]);
-    };
-    fetchCounts();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [associatedResources]);
-
-  useEffect(() => {
-    const resourcesLeft = checkIfAssociatedResourceLeft(
-      resourceCounts,
-      associatedResources,
-    );
-
-    setAssociatedResourceLeft(resourcesLeft);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [resourceCounts, associatedResources]);
-
   let startColumnComponent = null;
 
   const headerActions = (
@@ -211,126 +121,6 @@ const ColumnWraper = ({ defaultColumn = 'list', namespaced = false }) => {
       <Button onClick={() => handleResourceDelete({})} design="Transparent">
         {t('common.buttons.delete-module')}
       </Button>
-      {createPortal(
-        <DeleteMessageBox
-          disableDeleteButton={
-            associatedResourceLeft ? !allowForceDelete : false
-          }
-          customDeleteText={
-            associatedResourceLeft && allowForceDelete
-              ? 'common.buttons.cascade-delete'
-              : null
-          }
-          cancelFn={() => {
-            setAllowForceDelete(false);
-          }}
-          additionalDeleteInfo={
-            <>
-              <Text>
-                {t('kyma-modules.delete-module', {
-                  name: activeKymaModules[openedModuleIndex]?.name,
-                })}
-              </Text>
-              {associatedResources.length > 0 && (
-                <>
-                  <MessageStrip
-                    design="Information"
-                    hideCloseButton
-                    className="sap-margin-top-small"
-                  >
-                    {t('kyma-modules.associated-resources-warning')}
-                  </MessageStrip>
-                  <List
-                    headerText={t('kyma-modules.associated-resources')}
-                    mode="None"
-                    separators="All"
-                  >
-                    {associatedResources.map(assResource => {
-                      const resourceCount =
-                        resourceCounts[
-                          `${assResource.kind}-${assResource.group}-${assResource.version}`
-                        ];
-
-                      return (
-                        <ListItemStandard
-                          onClick={e => {
-                            e.preventDefault();
-                            handleItemClick(
-                              assResource.kind,
-                              assResource.group,
-                              assResource.version,
-                              clusterUrl,
-                              getScope,
-                              namespaceUrl,
-                              navigate,
-                            );
-                          }}
-                          type="Active"
-                          key={`${assResource.kind}-${assResource.group}-${assResource.version}`}
-                          additionalText={
-                            (resourceCount === 0 ? '0' : resourceCount) ||
-                            t('common.headers.loading')
-                          }
-                        >
-                          {pluralize(assResource?.kind)}
-                        </ListItemStandard>
-                      );
-                    })}
-                  </List>
-                  {associatedResourceLeft && (
-                    <CheckBox
-                      checked={allowForceDelete}
-                      onChange={() => setAllowForceDelete(!allowForceDelete)}
-                      accessibleName={t('kyma-modules.cascade-delete')}
-                      text={t('kyma-modules.cascade-delete')}
-                      className="sap-margin-top-tiny"
-                    />
-                  )}
-
-                  {associatedResourceLeft && allowForceDelete && (
-                    <MessageStrip
-                      design="Critical"
-                      hideCloseButton
-                      className="sap-margin-y-small"
-                    >
-                      {t('kyma-modules.cascade-delete-warning')}
-                    </MessageStrip>
-                  )}
-                </>
-              )}
-            </>
-          }
-          resourceTitle={activeKymaModules[openedModuleIndex]?.name}
-          deleteFn={() => {
-            if (allowForceDelete && forceDeleteUrls.length > 0) {
-              deleteAssociatedResources(
-                deleteResourceMutation,
-                forceDeleteUrls,
-              );
-            }
-            activeKymaModules.splice(openedModuleIndex, 1);
-            setKymaResourceState({
-              ...kymaResource,
-              spec: {
-                ...kymaResource.spec,
-                modules: activeKymaModules,
-              },
-            });
-            handleModuleUninstall();
-            setInitialUnchangedResource(cloneDeep(kymaResourceState));
-            setLayoutColumn({
-              layout: 'OneColumn',
-              startColumn: null,
-              midColumn: null,
-              endColumn: null,
-            });
-            if (allowForceDelete && forceDeleteUrls.length > 0) {
-              deleteCrResources(deleteResourceMutation, crUrls);
-            }
-          }}
-        />,
-        document.body,
-      )}
     </div>
   );
 
@@ -353,20 +143,14 @@ const ColumnWraper = ({ defaultColumn = 'list', namespaced = false }) => {
   } else {
     startColumnComponent = (
       <KymaModulesList
-        DeleteMessageBox={DeleteMessageBox}
-        handleResourceDelete={handleResourceDelete}
-        handleModuleUninstall={handleModuleUninstall}
-        setKymaResourceState={setKymaResourceState}
-        setInitialUnchangedResource={setInitialUnchangedResource}
         resourceName={kymaResource?.metadata?.name}
         resourceUrl={resourceUrl}
         kymaResource={kymaResource}
         kymaResourceLoading={kymaResourceLoading}
-        kymaResourceState={kymaResourceState}
         selectedModules={activeKymaModules}
-        setOpenedModuleIndex={setOpenedModuleIndex}
-        detailsOpen={detailsOpen}
         namespaced={namespaced}
+        setOpenedModuleIndex={setOpenedModuleIndex}
+        handleResourceDelete={handleResourceDelete}
       />
     );
   }
@@ -416,23 +200,44 @@ const ColumnWraper = ({ defaultColumn = 'list', namespaced = false }) => {
   );
 
   return (
-    <FlexibleColumnLayout
-      style={{ height: '100%' }}
-      layout={layoutState?.layout || 'OneColumn'}
-      startColumn={<div className="column-content">{startColumnComponent}</div>}
-      midColumn={
-        <>
-          {!layoutState?.showCreate &&
-            (defaultColumn !== 'details' || layout) && (
-              <div className="column-content">{detailsMidColumn}</div>
-            )}
-          {!layoutState?.midColumn &&
-            (defaultColumn !== 'details' || layout) && (
-              <div className="column-content">{createMidColumn}</div>
-            )}
-        </>
-      }
-    />
+    <>
+      {createPortal(
+        <ModulesDeleteBox
+          DeleteMessageBox={DeleteMessageBox}
+          selectedModules={activeKymaModules}
+          chosenModuleIndex={openedModuleIndex}
+          kymaResource={kymaResource}
+          kymaResourceState={kymaResourceState}
+          moduleTemplates={moduleTemplates}
+          detailsOpen={detailsOpen}
+          setKymaResourceState={setKymaResourceState}
+          setInitialUnchangedResource={setInitialUnchangedResource}
+          setChosenModuleIndex={setOpenedModuleIndex}
+          handleModuleUninstall={handleModuleUninstall}
+          setLayoutColumn={setLayoutColumn}
+        />,
+        document.body,
+      )}
+      <FlexibleColumnLayout
+        style={{ height: '100%' }}
+        layout={layoutState?.layout || 'OneColumn'}
+        startColumn={
+          <div className="column-content">{startColumnComponent}</div>
+        }
+        midColumn={
+          <>
+            {!layoutState?.showCreate &&
+              (defaultColumn !== 'details' || layout) && (
+                <div className="column-content">{detailsMidColumn}</div>
+              )}
+            {!layoutState?.midColumn &&
+              (defaultColumn !== 'details' || layout) && (
+                <div className="column-content">{createMidColumn}</div>
+              )}
+          </>
+        }
+      />
+    </>
   );
 };
